Replace libb2 with blake2
Currently, Sorbet uses blake2 for wasm builds and libb2 for other builds.
Originally, this was because libb2 is heavily optimized but not portable. Since
that time, however, libb2 has been extended with optimized builds for common
architectures. That means that it's now possible to use the same library
everywhere. x86/arm platforms still benefit from optimized vectorized
implementations, and other platforms can still use the portable fallback. This
change simplifies and slightly speeds up the build, while removing one of the
blockers to Linux AArch64 support as described in
https://github.com/sorbet/sorbet/issues/4119.